fix: remove telegram package
